pyproject.toml
[project]
name = "research-assistant"
version = "0.1.0"
description = "LangGraph-based deep research assistant"
requires-python = ">=3.10"
dependencies = [
    # LangGraph ecosystem
    "langgraph>=0.2.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",

    # LLM
    "ollama>=0.1.0",

    # Search & Scraping
    "duckduckgo-search>=4.0.0",
    "beautifulsoup4>=4.12.0",
    "newspaper3k>=0.2.8",
    "requests>=2.31.0",
    "lxml[html_clean]>=4.9.0",

    # Vector DB
    "chromadb>=0.4.0",

    # Embeddings
    "sentence-transformers>=2.2.0",

    # HTML Generation
    "jinja2>=3.1.0",
    "markdown>=3.5.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=5.0.0",
    "pytest-timeout>=2.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.12.0",
    "jupyter>=1.0.0",
]

[project.scripts]
research = "research_assistant.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
